api/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
anyio==4.3.0
    # via httpx
asgiref==3.7.2
    # via django
blessed==1.20.0
    # via curtsies
bpython==0.24
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
curtsies==0.4.2
    # via bpython
cwcwidth==0.1.9
    # via
    #   bpython
    #   curtsies
django==5.0.2
    # via djangorestframework
djangorestframework==3.14.0
exceptiongroup==1.2.0
    # via
    #   anyio
    #   pytest
greenlet==3.0.3
    # via
    #   bpython
    #   playwright
h11==0.14.0
    # via httpcore
httpcore==1.0.4
    # via httpx
httpx==0.27.0
idna==3.6
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.0.0
    # via pytest
packaging==23.2
    # via pytest
playwright==1.41.2
    # via pytest-playwright
pluggy==1.4.0
    # via pytest
pyee==11.0.1
    # via playwright
pygments==2.17.2
    # via bpython
pytest==8.0.1
    # via
    #   pytest-base-url
    #   pytest-django
    #   pytest-env
    #   pytest-playwright
pytest-base-url==2.1.0
    # via pytest-playwright
pytest-django==4.8.0
pytest-env==1.1.3
pytest-playwright==0.4.4
python-slugify==8.0.4
    # via pytest-playwright
pytz==2024.1
    # via djangorestframework
pyxdg==0.28
    # via bpython
requests==2.31.0
    # via
    #   bpython
    #   pytest-base-url
ruff==0.2.2
six==1.16.0
    # via blessed
sniffio==1.3.0
    # via
    #   anyio
    #   httpx
sqlparse==0.4.4
    # via django
strip-ansi==0.1.1
text-unidecode==1.3
    # via python-slugify
tomli==2.0.1
    # via
    #   pytest
    #   pytest-env
typing-extensions==4.9.0
    # via
    #   anyio
    #   asgiref
    #   pyee
urllib3==2.2.1
    # via requests
wcwidth==0.2.13
    # via blessed
